+++ legacy/Modules/PlazaModule.cs
@@ -12,5 +12,19 @@ namespace Orikivo.Modules
         [Summary("Get the status of each current shop.")]
         public async Task StoreCollectionResponseAsync()
             => await ModuleManager.TryExecute(Context.Channel, PlazaService.CheckStoresAsync(Context));
+
+        // Submodules
+        [Name("Pad")]
+        [Summary("Your own place of hope.")]
+        [DontAutoLoad]
+        public class CenterModule : ModuleBase<OrikivoCommandContext>
+        {
+            [Command("storage"), Alias("inventory", "inv", "stg")]
+            [Summary("Look at all of the currently owned items you have.")]
+            public async Task CheckStorageAsync()
+            {
+
+            }
+        }    
     }
 }
